refactor: deep simplify — 11 fixes from 3-agent review
- truncate: guard max=0 edge case (was usize underflow)
- format_age: clamp negative to 0 for future dates, extract format_age_from
- mode_hints: return &'static str instead of allocating String per frame
- Theme: store in App from config instead of hardcoding "nord" every frame
- PostbloxClient::new: return Result instead of panicking on build failure
- WsEvent::TrustChanged: approved_count i32 → i64 (no silent truncation)
- compose: also check lowercase "re: " to match MCP behavior
- inbox_list: make logical_selected() pub, remove wrapper
- search: remove no-op Layout::split, use block.inner directly
- app: remove redundant explicit drop(_guard)
- Delete unused event.rs (untested dead code)

+++ src/tui/ws.rs
@@ -18,7 +18,7 @@ pub enum WsEvent {
     TrustChanged {
         inbox_id: Uuid,
         new_mode: String,
-        approved_count: i32,
+        approved_count: i64,
     },
     Connected,
     Disconnected,
@@ -139,7 +139,7 @@ fn parse_ws_event(text: &str) -> Option<WsEvent> {
         "trust.changed" => {
             let inbox_id = parse_uuid(&msg.data, "inbox_id")?;
             let new_mode = msg.data["new_mode"].as_str()?.to_string();
-            let approved_count = msg.data["approved_count"].as_i64()? as i32;
+            let approved_count = msg.data["approved_count"].as_i64()?;
             Some(WsEvent::TrustChanged {
                 inbox_id,
                 new_mode,
